Update ExerciseDetailsPage.jsx
Added logic to fetch the exercise. If the exercise is passed through location.state then I would use that info. If the user enters through the url, it'll fetch the exercise.
Also, I'm saving the exercise and the videos for that exercise in sessionStorage so that the user does not call the API unnecessarily.